Add other persistent actors to refactor
