+++ persistence/src/main/scala/hmda/persistence/processing/HmdaFileParser.scala
@@ -6,7 +6,7 @@ import hmda.model.fi.ts.TransmittalSheet
 import hmda.parser.fi.lar.LarCsvParser
 import hmda.parser.fi.ts.TsCsvParser
 import hmda.persistence.CommonMessages._
-import hmda.persistence.LocalEventPublisher
+import hmda.persistence.{ HmdaPersistentActor, LocalEventPublisher }
 import hmda.persistence.processing.HmdaQuery._
 import hmda.persistence.processing.HmdaRawFile.LineAdded
 
@@ -42,7 +42,7 @@ object HmdaFileParser {
 
 }
 
-class HmdaFileParser(submissionId: String) extends HmdaPersistentActor(submissionId) with LocalEventPublisher {
+class HmdaFileParser(submissionId: String) extends HmdaPersistentActor with LocalEventPublisher {
 
   import HmdaFileParser._
